producer/requirements.txt
aio-pika==9.5.4
aiormq==6.8.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.6
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
minio==7.2.14
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pamqp==3.3.0
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
uvicorn[standard]
aiohttp
python-multipart